chore: moved/renamed root api routes
